+++ website/src/pages/_app.tsx
@@ -1,12 +1,9 @@
 import '~/styles/globals.css'
 import type { AppProps } from 'next/app'
-import Script from 'next/script'
 
 import { jetBrainsMono, manrope } from '~/styles/fonts'
 import { useRef } from 'react'
 import cx from 'classnames'
-
-const isProd = process.env.NODE_ENV === 'production'
 
 export default function App({ Component, pageProps }: AppProps) {
   const ref = useRef<HTMLElement>(null)
@@ -19,21 +16,6 @@ export default function App({ Component, pageProps }: AppProps) {
           font-family: ${manrope.style.fontFamily};
         }
       `}</style>
-
-      {isProd && (
-        <>
-          <Script strategy="lazyOnload" src={`https://www.googletagmanager.com/gtag/js?id=G-Y678D4CC1J`} />
-          <Script id="gtm" strategy="lazyOnload">
-            {`
-        window.dataLayer = window.dataLayer || [];
-        function gtag(){dataLayer.push(arguments);}
-        gtag('js', new Date());
-        
-        gtag('config', 'G-Y678D4CC1J');
-      `}
-          </Script>
-        </>
-      )}
 
       <main className={cx(jetBrainsMono.variable, manrope.variable)} ref={ref}>
         <Component {...pageProps} />